+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @naive-potato-ui/ui
 
+## 1.0.25
+
+### Patch Changes
+
+- table info custom render
+
+- Updated dependencies []:
+  - @naive-potato-ui/curd-table@1.0.22
+
 ## 1.0.24
 
 ### Patch Changes